Update monitoring workflow with new jobs and amounts (#3916)
* Update monitoring workflow with new jobs and amounts
