+++ .github/workflows/monitoring-limit-geo-e2e-tests.yml
@@ -3,7 +3,7 @@ name: Synthetic Limit Orders Geo Monitoring Frontend tests
 on:
   workflow_dispatch:
   schedule:
-    - cron: "0 * * * *"
+    - cron: "4 * * * *"
 
 jobs:
   fe-trade-eu-tests:
@@ -37,13 +37,54 @@ jobs:
           USE_TEST_PROXY: "use"
         run: |
           cd packages/web
-          npx playwright test monitoring --timeout 180000
+          npx playwright test monitoring.market --timeout 180000
       - name: upload monitoring test results
         if: failure()
         id: monitoring-test-results
         uses: actions/upload-artifact@v4
         with:
-          name: eu-trade-test-results
+          name: eu-trade-test-results-${{ github.run_id }}
+          path: packages/web/playwright-report
+
+  fe-limit-eu-tests:
+    timeout-minutes: 12
+    name: prod-fe-trade-eu-tests
+    needs: fe-trade-eu-tests
+    runs-on: macos-latest
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
+      - name: Cache dependencies
+        uses: actions/cache@v4
+        with:
+          path: "**/node_modules"
+          key: ${{ runner.OS }}-20.x-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.OS }}-20.x-
+      - name: Install Playwright
+        run: |
+          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+      - name: Run Limit tests in EU
+        env:
+          BASE_URL: "https://app.osmosis.zone"
+          TEST_PROXY: "http://138.68.112.16:8888"
+          TEST_PROXY_USERNAME: ${{ secrets.TEST_PROXY_USERNAME }}
+          TEST_PROXY_PASSWORD: ${{ secrets.TEST_PROXY_PASSWORD }}
+          PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY_1 }}
+          USE_TEST_PROXY: "use"
+        run: |
+          cd packages/web
+          npx playwright test monitoring.limit --timeout 180000
+      - name: upload monitoring test results
+        if: failure()
+        id: monitoring-test-results
+        uses: actions/upload-artifact@v4
+        with:
+          name: eu-limit-test-results-${{ github.run_id }}
           path: packages/web/playwright-report
 
   fe-trade-sg-tests:
@@ -77,13 +118,13 @@ jobs:
           USE_TEST_PROXY: "use"
         run: |
           cd packages/web
-          npx playwright test monitoring --timeout 180000
+          npx playwright test monitoring.market --timeout 180000
       - name: upload monitoring test results
         if: failure()
         id: monitoring-test-results
         uses: actions/upload-artifact@v4
         with:
-          name: sg-trade-test-results
+          name: sg-trade-test-results-${{ github.run_id }}
           path: packages/web/playwright-report
 
   fe-trade-us-tests:
@@ -114,18 +155,64 @@ jobs:
           PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY_3 }}
         run: |
           cd packages/web
-          npx playwright test monitoring --timeout 180000
+          npx playwright test monitoring.market --timeout 180000
       - name: upload monitoring test results
         if: failure()
         id: monitoring-test-results
         uses: actions/upload-artifact@v4
         with:
-          name: us-trade-test-results
+          name: us-trade-test-results-${{ github.run_id }}
+          path: packages/web/playwright-report
+
+  fe-limit-us-tests:
+    timeout-minutes: 10
+    needs: fe-trade-us-tests
+    name: prod-fe-limit-us-tests
+    runs-on: macos-latest
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
+      - name: Cache dependencies
+        uses: actions/cache@v4
+        with:
+          path: "**/node_modules"
+          key: ${{ runner.OS }}-20.x-${{ hashFiles('**/yarn.lock') }}
+          restore-keys: |
+            ${{ runner.OS }}-20.x-
+      - name: Install Playwright
+        run: |
+          yarn --cwd packages/web install --frozen-lockfile && npx playwright install --with-deps chromium
+      - name: Run Limit tests in US
+        id: run-us
+        env:
+          BASE_URL: "https://app.osmosis.zone"
+          PRIVATE_KEY: ${{ secrets.TEST_PRIVATE_KEY_3 }}
+        run: |
+          cd packages/web
+          npx playwright test monitoring.limit --timeout 180000
+      - name: upload monitoring test results
+        if: failure()
+        id: monitoring-test-results
+        uses: actions/upload-artifact@v4
+        with:
+          name: us-limit-test-results-${{ github.run_id }}
           path: packages/web/playwright-report
 
   fe-bot-alert:
     runs-on: ubuntu-latest
-    needs: [fe-trade-eu-tests, fe-trade-us-tests, fe-trade-sg-tests]
+    needs:
+      [
+        fe-trade-eu-tests,
+        fe-trade-us-tests,
+        fe-trade-sg-tests,
+        fe-trade-us-tests,
+        fe-limit-us-tests,
+        fe-limit-eu-tests,
+      ]
     if: failure()
     steps:
       - name: Send Slack alert if test fails
@@ -159,7 +246,15 @@ jobs:
     runs-on: ubuntu-latest
     if: always()
     needs:
-      [fe-trade-eu-tests, fe-trade-us-tests, fe-trade-sg-tests, fe-bot-alert]
+      [
+        fe-trade-eu-tests,
+        fe-trade-us-tests,
+        fe-trade-sg-tests,
+        fe-bot-alert,
+        fe-trade-us-tests,
+        fe-limit-us-tests,
+        fe-limit-eu-tests,
+      ]
     steps:
       - name: Delete Previous deployments
         uses: actions/github-script@v7
