+++ .github/workflows/prod-frontend-e2e-tests.yml
@@ -56,7 +56,7 @@ jobs:
         env:
           BASE_URL: "https://app.osmosis.zone"
           PRIVATE_KEY: ${{ secrets.PRIVATE_KEY }}
-          USE_TRADE: "use"
+          WALLET_ID: ${{ secrets.TEST_WALLET_ID }}
         run: |
           cd packages/web
           npx playwright test -g "Test Swap feature"
@@ -65,7 +65,7 @@ jobs:
         id: e2e-test-results
         uses: actions/upload-artifact@v4
         with:
-          name: main-e2e-test-results
+          name: main-e2e-test-results-${{ github.run_id }}
           path: packages/web/playwright-report
       - name: Send Slack alert if test fails
         id: slack
