+++ packages/web/e2e/tests/monitoring.market.wallet.spec.ts
@@ -2,7 +2,6 @@
 import { type BrowserContext, chromium, expect, test } from "@playwright/test";
 import process from "process";
 
-import { TransactionsPage } from "~/e2e/pages/transactions-page";
 import { TestConfig } from "~/e2e/test-config";
 import { UnzipExtension } from "~/e2e/unzip-extension";
 
@@ -44,58 +43,13 @@ test.describe("Test Filled Limit Order feature", () => {
     await context.close();
   });
 
-  test("User should be able to limit sell OSMO", async () => {
-    await tradePage.goto();
-    await tradePage.openSellTab();
-    await tradePage.openLimit();
-    await tradePage.selectAsset("OSMO");
-    await tradePage.enterAmount("1.01");
-    await tradePage.setLimitPriceChange("Market");
-    const { msgContentAmount } = await tradePage.sellAndGetWalletMsg(
-      context,
-      true
-    );
-    expect(msgContentAmount).toBeTruthy();
-    // now this is converted from USDC
-    expect(msgContentAmount).toContain("place_limit");
-    expect(msgContentAmount).toContain('"order_direction": "ask"');
-    await tradePage.isTransactionSuccesful();
-    await tradePage.getTransactionUrl();
-  });
-
-  test("User should be able to limit buy OSMO", async () => {
-    await tradePage.goto();
-    await tradePage.openBuyTab();
-    await tradePage.openLimit();
-    await tradePage.selectAsset("OSMO");
-    await tradePage.enterAmount("1.01");
-    await tradePage.setLimitPriceChange("Market");
-    const limitPrice = Number(await tradePage.getLimitPrice());
-    const highLimitPrice = (limitPrice * 1.07).toFixed(4);
-    await tradePage.setLimitPrice(String(highLimitPrice));
-    const { msgContentAmount } = await tradePage.buyAndGetWalletMsg(
-      context,
-      true
-    );
-    expect(msgContentAmount).toBeTruthy();
-    expect(msgContentAmount).toContain('"quantity": "1010000"');
-    expect(msgContentAmount).toContain("place_limit");
-    expect(msgContentAmount).toContain('"order_direction": "bid"');
-    await tradePage.isTransactionSuccesful();
-    await tradePage.getTransactionUrl();
-    await tradePage.gotoOrdersHistory(30);
-    const p = context.pages()[0];
-    const trxPage = new TransactionsPage(p);
-    await trxPage.isFilledByLimitPrice(highLimitPrice);
-  });
-
   // biome-ignore lint/complexity/noForEach: <explanation>
   [{ name: "WBTC" }, { name: "OSMO" }].forEach(({ name }) => {
     test(`User should be able to Market Buy ${name}`, async () => {
       await tradePage.goto();
       await tradePage.openBuyTab();
       await tradePage.selectAsset(name);
-      await tradePage.enterAmount("0.52");
+      await tradePage.enterAmount("0.75");
       const { msgContentAmount } = await tradePage.buyAndGetWalletMsg(context);
       expect(msgContentAmount).toBeTruthy();
       expect(msgContentAmount).toContain("type: osmosis/poolmanager/");
@@ -105,11 +59,11 @@ test.describe("Test Filled Limit Order feature", () => {
   });
 
   // unwrapped market sell tests just in case this affects anything.
-  test.skip("User should be able to Market Sell WBTC", async () => {
+  test("User should be able to Market Sell WBTC", async () => {
     await tradePage.goto();
     await tradePage.openSellTab();
     await tradePage.selectAsset("WBTC");
-    await tradePage.enterAmount("0.51");
+    await tradePage.enterAmount("0.74");
     await tradePage.isSufficientBalanceForTrade();
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.sellAndGetWalletMsg(context);
@@ -123,7 +77,7 @@ test.describe("Test Filled Limit Order feature", () => {
     await tradePage.goto();
     await tradePage.openSellTab();
     await tradePage.selectAsset("OSMO");
-    await tradePage.enterAmount("0.51");
+    await tradePage.enterAmount("0.74");
     await tradePage.isSufficientBalanceForTrade();
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.sellAndGetWalletMsg(context);
