+++ packages/web/e2e/tests/stables.wallet.spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import { BrowserContext, chromium, expect, test } from "@playwright/test";
+import { type BrowserContext, chromium, expect, test } from "@playwright/test";
 import process from "process";
 
 import { TradePage } from "~/e2e/pages/trade-page";
@@ -12,12 +12,13 @@ test.describe("Test Swap Stables feature", () => {
   let context: BrowserContext;
   const privateKey = process.env.PRIVATE_KEY ?? "private_key";
   let tradePage: TradePage;
-  let USDC =
+  const USDC =
     "ibc/498A0751C798A0D9A389AA3691123DADA57DAA4FE165D5C75894505B876BA6E4";
-  let USDCa =
+  const USDCa =
     "ibc/D189335C6E4A68B513C10AB227BF1C1D38C746766278BA3EEB4FB14124F1D858";
-  let allUSDT =
+  const allUSDT =
     "factory/osmo1em6xs47hd82806f5cxgyufguxrrc7l0aqx7nzzptjuqgswczk8csavdxek/alloyed/allUSDT";
+  const swapAmount = "0.25";
 
   test.beforeAll(async () => {
     const pathToExtension = new UnzipExtension().getPathToExtension();
@@ -52,12 +53,12 @@ test.describe("Test Swap Stables feature", () => {
   test("User should be able to swap USDC to USDC.eth.axl", async () => {
     await tradePage.goto();
     await tradePage.selectPair("USDC", "USDC.eth.axl");
-    await tradePage.enterAmount("0.1");
+    await tradePage.enterAmount(swapAmount);
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.swapAndGetWalletMsg(context);
     expect(msgContentAmount).toBeTruthy();
-    expect(msgContentAmount).toContain("denom: " + USDC);
-    expect(msgContentAmount).toContain("token_out_denom: " + USDCa);
+    expect(msgContentAmount).toContain(`denom: ${USDC}`);
+    expect(msgContentAmount).toContain(`token_out_denom: ${USDCa}`);
     await tradePage.isTransactionSuccesful();
     await tradePage.getTransactionUrl();
   });
@@ -65,12 +66,12 @@ test.describe("Test Swap Stables feature", () => {
   test("User should be able to swap USDC.eth.axl to USDC", async () => {
     await tradePage.goto();
     await tradePage.selectPair("USDC.eth.axl", "USDC");
-    await tradePage.enterAmount("0.1");
+    await tradePage.enterAmount(swapAmount);
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.swapAndGetWalletMsg(context);
     expect(msgContentAmount).toBeTruthy();
-    expect(msgContentAmount).toContain("denom: " + USDCa);
-    expect(msgContentAmount).toContain("token_out_denom: " + USDC);
+    expect(msgContentAmount).toContain(`denom: ${USDCa}`);
+    expect(msgContentAmount).toContain(`token_out_denom: ${USDC}`);
     await tradePage.isTransactionSuccesful();
     await tradePage.getTransactionUrl();
   });
@@ -78,7 +79,7 @@ test.describe("Test Swap Stables feature", () => {
   test("User should be able to swap USDT to USDC", async () => {
     await tradePage.goto();
     await tradePage.selectPair("USDT", "USDC");
-    await tradePage.enterAmount("0.1");
+    await tradePage.enterAmount(swapAmount);
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.swapAndGetWalletMsg(context);
     expect(msgContentAmount).toBeTruthy();
@@ -91,11 +92,11 @@ test.describe("Test Swap Stables feature", () => {
   test("User should be able to swap USDC to USDT", async () => {
     await tradePage.goto();
     await tradePage.selectPair("USDC", "USDT");
-    await tradePage.enterAmount("0.1");
+    await tradePage.enterAmount(swapAmount);
     await tradePage.showSwapInfo();
     const { msgContentAmount } = await tradePage.swapAndGetWalletMsg(context);
     expect(msgContentAmount).toBeTruthy();
-    expect(msgContentAmount).toContain("denom: " + USDC);
+    expect(msgContentAmount).toContain(`denom: ${USDC}`);
     expect(msgContentAmount).toContain(allUSDT);
     await tradePage.isTransactionSuccesful();
     await tradePage.getTransactionUrl();
